chore: extract MCDARangeControls into separate component
